// src/modules/tenantUser/enum/auth.enum.ts
export enum AuthEnum {
  SOMETHING_WENT_WRONG = 'Something went wrong',
  USER_NOT_FOUND = 'User not found',
  REQUIRED_FIELDS = 'Missing required field',
  INVALID_CREDENTIALS = 'Invalid credentials',
  NOT_ALLOWED_TO_LOGIN = 'You are not allowed to login here',
  INVALID_ACCESS_TOKEN = 'Invalid access token',
  NOT_ENOUGH_RIGHTS = 'Not enough rights',
  UNAUTHORIZED = 'Unauthorized',
  OTP_NOT_VERIFIED = 'OTP is not correct',
  OTP_VERIFIED = 'OTP matched successfully',
  OTP_SEND_SUCCESS = 'OTP sended successfully',
  LOGIN_SUCCESS = 'Login success',
  REGISTER_SUCCESS = 'Register success',
  RESET_SUCCESS = 'Password Reset Successful',
  INVALID_EMAIL = 'Invalid Email Address',
  EMAIL_EXISTS = 'Email already exists',
  PASSWORD_MIN_LENGTH = 'Password must be at least 6 characters',
  INTEGER_ERROR = 'Value must be at an integer',
  INVALID_TOKEN = 'Invalid Token',
  INVALID_DATE_FORMAT = 'Invalid date format',
  INVALID_GENDER = 'Invalid gender value',
  INVALID_ROLE = 'Role Not Found',
  INVALID_STATUS = 'Status Not Found',
}